Portable java file, version 0.1.1
This should work as a cross-platform package for my project. Will update this every time there is a "major or minor" update to the software (ie. new features and/or bugfixes). Tested on two systems running windows 10 and a single Ubuntu 18.04 virtual machine. I do appreciate feedback if this package isn't working on your system.